fix: linter errors

+++ pkg/core/src/tests/test-helpers/LiquidityHelper.sol
@@ -10,6 +10,7 @@ import { ISwapRouter } from "@uniswap/v3-periphery/contracts/interfaces/ISwapRou
 
 interface WstETHInterface {
     function unwrap(uint256 _wstETHAmount) external returns (uint256);
+
     function wrap(uint256 _stETHAmount) external returns (uint256);
 }
 
@@ -18,9 +19,8 @@ interface StETHInterface {
 }
 
 interface CTokenInterface {
-    function mint(uint mintAmount) external returns (uint);
+    function mint(uint256 mintAmount) external returns (uint256);
 }
-
 
 contract LiquidityHelper {
     using SafeERC20 for ERC20;
@@ -34,10 +34,15 @@ contract LiquidityHelper {
         }
     }
 
-    function swap(address tokenIn, address tokenOut, uint256 amountIn, address recipient) public returns (uint256) {
+    function swap(
+        address tokenIn,
+        address tokenOut,
+        uint256 amountIn,
+        address recipient
+    ) public returns (uint256) {
         uint256 amountOut = 0;
         if (tokenOut == Assets.WSTETH) {
-            uint256 stETH = StETHInterface(Assets.STETH).submit{value: 10 ether}(address(0));
+            uint256 stETH = StETHInterface(Assets.STETH).submit{ value: 10 ether }(address(0));
             ERC20(Assets.STETH).safeApprove(Assets.WSTETH, stETH);
             amountOut = WstETHInterface(Assets.WSTETH).wrap(stETH);
             emit Swapped(tokenIn, tokenOut, amountIn, amountOut);
@@ -49,19 +54,18 @@ contract LiquidityHelper {
             emit Swapped(tokenIn, tokenOut, amountIn, amountOut);
             return amountOut;
         }
-        
+
         // approve router to spend tokenIn
         ERC20(tokenIn).safeApprove(Assets.UNISWAP_ROUTER, amountIn);
-        ISwapRouter.ExactInputSingleParams memory params =
-            ISwapRouter.ExactInputSingleParams({
-                tokenIn: tokenIn,
-                tokenOut: tokenOut,
-                fee: UNI_POOL_FEE,
-                recipient: recipient,
-                deadline: block.timestamp,
-                amountIn: amountIn,
-                amountOutMinimum: 0,
-                sqrtPriceLimitX96: 0 // set to be 0 to ensure we swap our exact input amount
+        ISwapRouter.ExactInputSingleParams memory params = ISwapRouter.ExactInputSingleParams({
+            tokenIn: tokenIn,
+            tokenOut: tokenOut,
+            fee: UNI_POOL_FEE,
+            recipient: recipient,
+            deadline: block.timestamp,
+            amountIn: amountIn,
+            amountOutMinimum: 0,
+            sqrtPriceLimitX96: 0 // set to be 0 to ensure we swap our exact input amount
         });
         amountOut = ISwapRouter(Assets.UNISWAP_ROUTER).exactInputSingle(params); // executes the swap
         emit Swapped(tokenIn, tokenOut, amountIn, amountOut);
@@ -69,5 +73,4 @@ contract LiquidityHelper {
     }
 
     event Swapped(address tokenIn, address tokenOut, uint256 amountIn, uint256 amountOut);
-
 }
